requirements.txt
# Updated package versions (tested and working)
tensorflow>=2.13.0,<3.0.0
numpy>=1.24.0,<3.0.0
matplotlib>=3.7.0
pillow>=10.0.0
opencv-python>=4.8.0
jupyter>=1.0.0
flask>=2.3.0
keras>=3.0.0
seaborn>=0.12.0
scikit-learn>=1.3.0
